+++ src/OpenTelemetry.Exporter.OneCollector/OneCollectorExporterSerializationExceptionStackTraceHandlingType.cs
@@ -1,0 +1,35 @@
+// <copyright file="OneCollectorExporterSerializationExceptionStackTraceHandlingType.cs" company="OpenTelemetry Authors">
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// </copyright>
+
+namespace OpenTelemetry.Exporter.OneCollector;
+
+/// <summary>
+/// Describes the OneCollector serialization behavior to use when writing the
+/// stack trace for telemetry containing an <see cref="Exception"/> instance.
+/// </summary>
+public enum OneCollectorExporterSerializationExceptionStackTraceHandlingType
+{
+    /// <summary>
+    /// Exception stack traces will be ignored when serializing.
+    /// </summary>
+    Ignore,
+
+    /// <summary>
+    /// Exception stack traces will be included when serializing by calling <see
+    /// cref="Exception.ToString"/>.
+    /// </summary>
+    IncludeAsString,
+}