+++ src/OpenTelemetry.Exporter.OneCollector/CHANGELOG.md
@@ -6,9 +6,13 @@
   syntax.
   ([#1073](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1073))
 
-* Added support for sending common schema `dt` (Distributed Tracing) extensions
+* Added support for sending common schema `dt` (Distributed Tracing) extension
   when trace context is present on `LogRecord`s.
   ([#1073](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1073))
+
+* Added support for sending common schema `ex` (Exception) extension when
+  exception is present on `LogRecord`s.
+  ([#1082](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1082))
 
 ## 0.1.0-alpha.2